chore: refactor functions to helper class
